move LoadOrBuildDag & LoadFrozenData into its own file
